requirements.txt
ansible==8.5.0
ansible-core==2.15.5
awscli==1.29.22
boto==2.49.0
boto3==1.28.22
botocore==1.31.22
cryptography==36.0.1
colorama==0.4.3
dnspython==2.2.0
docker==5.0.3
docker-compose==1.29.2
netaddr==0.8.0
PyYAML==5.4.1
python-jenkins==1.7.0
lxml==4.8.0
Jinja2==3.0.3
pysocks==1.7.1
pypsrp==0.8.1
requests_credssp==2.0.0
tox==3.24.5
pyopenssl==22.0.0
certifi==2022.06.15
